[autoscaling] Conditionally run the scale up and down test
The scale-up and scale downs tests should not run unless the cpu_high
or cpu_low alarm is triggered.
If the alarm is not triggered, then the scale up/down test is always going
to fail, so we don't need to waste 500s waiting for it to fail.

The test will be marked as skipped, but could also be marked as failed
instead.