[cloukitty] Set some facts instead of env vars in the command task
Since we need to split this up, using env cars is not going to work,
since each task is it's own shell, and values don't transfer.

Instead of repeating the same lines each time, we can set some facts to
be used throughout.
In this case, it's just for the values that we know before hand.
The facts can be referenced in the shell tasks (or any string) using
jinja2 templating syntax.

The large shell script doesn't return any output until it's completed.
This might mean that there's some error, timeout or hung state and we
won't know until the task times out.