Make some bash to ansible conversions: split up some statements
Split out more statements into tasks

There are a bunch of vars we need between tasks, so we need to capture
the outputs somehow.
We can register the task result to capture the output for later use!

The suggestions from ansible-lint are more actionable at this point.

We can also start seeing errors in this case because the resources
already exist. (Oh, idempotency, where art thou?)
With ansible, we can wrap the tasks in a block/always to make sure the
cleanup tasks always get done.