Add failed_when and changed_when to tasks
This customises how the tasks are treated and allows to conditionally
deal with failures, i.e. some failures are okay, others depend on the
output of the task.

For shell tasks that get information, they can be set to changed_when:
false since they are never going to change anything

In some cases, commands with rc != 0 is okay, but fail by default,
failed_when lets you set the condition that causes a failed task.

If the pass/fail conditions are based on some value being in the stdout,
you can set failed_when based on stdout content