Skip importing from extra_vars if cifmw_basedir does not exist
This might be temperary, but needs to be investigated as to whether the vars file is needed.
One option is to copy all the relevant vars into fvt/ci/vars/ and import them directly into the playbooks
Another option is to copy the artefacts dir into the logs volume that gets mounted into the pods.